OBGM-47 Refactor ProductCatalog and productCatalogItem ExcelIporters
