+++ grails-app/services/org/pih/warehouse/product/ProductCatalogService.groovy
@@ -6,7 +6,7 @@
 * By using this software in any fashion, you are agreeing to be bound by
 * the terms of this license.
 * You must not remove this notice, or any other, from this software.
-**/ 
+**/
 package org.pih.warehouse.product
 
 class ProductCatalogService {
@@ -32,5 +32,17 @@ class ProductCatalogService {
         }
 
         return productCatalogItem
+    }
+
+    ProductCatalog createOrUpdateProductCatalog(Map params) {
+        ProductCatalog productCatalog = ProductCatalog.findByIdOrName(params.id, params.name)
+        if (!productCatalog) {
+            productCatalog = new ProductCatalog()
+        }
+        productCatalog.id = params.id
+        productCatalog.code = params.code
+        productCatalog.name = params.name
+        productCatalog.description = params.description
+        return productCatalog
     }
 }
